refactor: consolidate context imports and enhance context structure for better readability
