+++ apps/web/src/routes/kiwi/-reader/index.ts
@@ -1,9 +1,6 @@
-export * from "./book-context";
 export { default as ReaderContents } from "./contents";
 export * from "./flip-page-button";
 export { default as ReaderPageProgress } from "./page-progress";
 export { default as ReaderProvider } from "./provider";
-export * from "./settings-context";
+export * from "./contexts";
 export * from "./styles";
-export * from "./record-context";
-export * from "./reading-context";
